+++ packages/cli/src/utils/get-package-json.ts
