+++ packages/cli/src/utils/get-version.ts
